+++ atrium-api/src/app/bsky/feed/generator.rs
@@ -5,7 +5,7 @@
 pub struct Record {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub avatar: Option<crate::types::BlobRef>,
-    pub created_at: String,
+    pub created_at: crate::types::string::Datetime,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]