+++ atrium-api/CHANGELOG.md
@@ -21,8 +21,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - All Lexicon string fields with one of the following formats now have the corresponding
   dedicated type, instead of `String`:
   - `at-identifier` (`atrium_api::types::string::AtIdentifier`)
+  - `datetime` (`atrium_api::types::string::Datetime`)
   - `did` (`atrium_api::types::string::Did`)
   - `handle` (`atrium_api::types::string::Handle`)
+  - `nsid` (`atrium_api::types::string::Nsid`)
+  - `language` (`atrium_api::types::string::Language`)
 
 ## [0.16.0](https://github.com/sugyan/atrium/compare/atrium-api-v0.15.0...atrium-api-v0.16.0) - 2024-02-09